Intelligent Fuzzy Logic Irrigation System/requirements.txt
requests
python-dotenv
numpy
pytest